Replace Qt signal/slot mechanism by Observer pattern where possible
The signal/slot mechanism seems to sometimes break the definition of an interface requiring you to inherit from QObject class.
One of the caveats is that it can lead to the Dreaded Diamond issue.
All in all, currently it is a try to avoid using that mechanism where possible. Later it will become clearer was it a right decision or not.